+++ src/core/features.js
@@ -23,7 +23,7 @@ onix.service("$features", function() {
 	this.CANVAS = !!document.createElement("canvas").getContext;
 
 	// local storage
-	var locStor = true;
+	let locStor = true;
 
 	try {
 		window.localStorage;
@@ -47,4 +47,22 @@ onix.service("$features", function() {
 	 * @type {Boolean}
 	 */
 	this.MEDIA_QUERY = "matchMedia" in window && "matches" in window.matchMedia("(min-width: 500px)");
+
+	// mouse wheel event name
+	let mouseWheel = "DOMMouseScroll";
+
+	if ("onwheel" in window) {
+		mouseWheel = "wheel";
+	}
+	else if ("onmousewheel" in window) {
+		mouseWheel = "mousewheel";
+	}
+
+	/**
+	 * Event name for mouse wheel.
+	 *
+	 * @member $features
+	 * @type {String}
+	 */
+	this.MOUSE_WHEEL_EVENT_NAME = mouseWheel;
 });