+++ src/core/location.js
@@ -12,7 +12,7 @@ onix.service("$location", function() {
 	 * @member $location
 	 */
 	this.refresh = function() {
-		window.location.reload();
+		location.reload();
 	};
 
 	/**
@@ -65,26 +65,12 @@ onix.service("$location", function() {
 			let newURL = this.createSearchURL(obj);
 
 			if (newURL) {
-				window.location.search = newURL;
+				location.search = newURL;
 			}
 		}
 		else {
 			// read
-			let data = location.search;
-			let output = {};
-
-			if (data) {
-				data = data.replace("?", "");
-
-				data.split("&").forEach(item => {
-					let parts = item.split("=");
-					let name = decodeURIComponent(parts[0]);
-					
-					output[name] = decodeURIComponent(parts[1]);
-				});
-			}
-
-			return output;
+			return this.parseSearch();
 		}
 	};
 
@@ -95,6 +81,173 @@ onix.service("$location", function() {
 	 * @member $location
 	 */
 	this.get = function() {
-		return window.location.pathname + window.location.search;
+		return location.pathname + location.search;
+	};
+
+	/**
+	 * Decode value from URL.
+	 * 
+	 * @param  {String} value Input value
+	 * @return {String}
+	 * @member $location
+	 */
+	this.decodeSearchValue = function(value) {
+		return decodeURIComponent(value.replace(/\+/g, " "));
+	};
+
+	/**
+	 * Parse search part of the URL.
+	 * 
+	 * @param  {String} [query] Optinal query, default is location.search
+	 * @return {Object} Object with keys and values from the search
+	 * @member $location
+	 */
+	this.parseSearch = function(query) {
+		// read
+		query = query || location.search.substring(1);
+
+		let match;
+		let search = /([^&=]+)=?([^&]*)/g;
+		let output = {};
+
+		while (match = search.exec(query)) {
+			let key = this.decodeSearchValue(match[1]);
+			let value = this.decodeSearchValue(match[2]);
+
+			if (key in output) {
+				if (!Array.isArray(output[key])) {
+					output[key] = [output[key]];
+				}
+
+				output[key].push(value);
+			}
+			else {
+				output[key] = value;
+			}
+		}
+
+		return output;
+	};
+
+	/**
+	 * Parse URL to object.
+	 * 
+	 * @param {String} url Input URL
+	 * @param {Object} [optsArg] optional configuration
+	 * @param {Boolean} [optsArg.autoNumber = false] find number in string and convert it
+	 * @param {Object} [optsArg.hints = {}] { key name : convert operation }, operations: "json" value -> object, "number" -> value -> number, fn(value) -> value
+	 * @return {Object} parse url to object with keys like host, protocol etc.
+	 * @member $location
+	 */
+	this.parseURL = function(url, optsArg) {
+		let opts = {
+			autoNumber: false,
+			hints: {}
+		};
+
+		let obj = {
+			protocol: "",
+			host: "",
+			port: null,
+			path: "",
+			search: null,
+			hash: ""
+		};
+
+		for (let key in optsArg) {
+			opts[key] = optsArg[key];
+		}
+
+		url = (url || "").trim();
+
+		// protocol
+		let test = url.match(/([a-zA-Z0-9]+):\/\//);
+
+		if (test) {
+			obj.protocol = test[1];
+			url = url.replace(test[0], "");
+		}
+
+		// host
+		test = url.match(/^[^?:#\/]+/);
+
+		if (test) {
+			obj.host = test[0];
+			url = url.replace(obj.host, "");
+		}
+
+		// port
+		test = url.match(/^:([0-9]+)[\/?#]?/);
+
+		if (test) {
+			obj.port = parseFloat(test[1]);
+			url = url.replace(":" + test[1], "");
+		}
+
+		// path
+		test = url.match(/^[^?#]+/);
+
+		if (test) {
+			obj.path = test[0];
+			url = url.replace(obj.path, "");
+		}
+
+		// search
+		test = url.match(/\?([^#]+)/);
+
+		if (test) {
+			let searchObj = this.parseSearch(test[1]);
+
+			// update
+			Object.keys(searchObj).forEach(key => {
+				let value = searchObj[key];
+
+				if (key in opts.hints) {
+					let hintValue = opts.hints[key];
+
+					if (typeof hintValue === "string") {
+						switch (opts.hints[key]) {
+							case "json":
+								try {
+									searchObj[key] = JSON.parse(value);
+								}
+								catch (err) {
+									console.error(err);
+								}
+								break;
+
+							case "number":
+								searchObj[key] = parseFloat(value);
+								break;
+						}
+					}
+					else if (typeof hintValue === "function") {
+						searchObj[key] = hintValue(value);
+					}
+				}
+				else if (opts.autoNumber) {
+					let numTest = value.match(/^[-]?[0-9]+\.?[0-9e]*$/);
+
+					if (numTest) {
+						let num = parseFloat(numTest[0]);
+
+						searchObj[key] = isNaN(num) ? value : num;
+					}
+				}
+			});
+
+			obj.search = searchObj;
+
+			url = url.replace(test[0], "");
+		}
+
+		// hash
+		test = url.match(/#(.*)$/);
+
+		if (test) {
+			obj.hash = test[1];
+		}
+
+		return obj;
 	};
 });